+++ documentation/ChangeLog.md
@@ -1,6 +1,7 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Added on-the-fly parser errors in the editor (uses Go `oracle`).
  * Added some new Snippets: `for`, `foreach`, `switch`, and a few others.
  * Fixed NPE when invoking Open Definition in a file that is available in the Eclipse workspace under more than one path/location. (#167)
  * Fixed bug with Content Assist snippets using the `${word_selection}` variable.